pyproject.toml
[project]
name = "capstone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "argon2-cffi==25.1.0",
    "argon2-cffi-bindings==21.2.0",
    "arrow==1.3.0",
    "asttokens==3.0.0",
    "async-lru==2.0.5",
    "attrs==25.3.0",
    "babel==2.17.0",
    "beautifulsoup4==4.13.4",
    "bleach==6.2.0",
    "cachetools==5.5.2",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "comm==0.2.2",
    "debugpy==1.8.14",
    "decorator==5.2.1",
    "defusedxml==0.7.1",
    "et-xmlfile==2.0.0",
    "executing==2.2.0",
    "fastjsonschema==2.21.1",
    "fqdn==1.5.1",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "google-genai==1.26.0",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.73.1",
    "grpcio-status==1.73.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.29.5",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "isoduration==20.11.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "json5==0.12.0",
    "jsonpointer==3.0.0",
    "jsonschema==4.24.0",
    "jsonschema-specifications==2025.4.1",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "jupyter-events==0.12.0",
    "jupyter-lsp==2.2.5",
    "jupyter-server==2.16.0",
    "jupyter-server-terminals==0.5.3",
    "jupyterlab==4.4.4",
    "jupyterlab-pygments==0.3.0",
    "jupyterlab-server==2.27.3",
    "litellm>=1.75.5.post1",
    "markupsafe==3.0.2",
    "matplotlib>=3.10.6",
    "matplotlib-inline==0.1.7",
    "mistune==3.1.3",
    "nbclient==0.10.2",
    "nbconvert==7.16.6",
    "nbformat==5.10.4",
    "nest-asyncio==1.6.0",
    "notebook==7.4.4",
    "notebook-shim==0.2.4",
    "numpy==2.3.1",
    "openpyxl==3.1.5",
    "overrides==7.7.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pandocfilters==1.5.1",
    "parso==0.8.4",
    "pdf2image>=1.17.0",
    "pexpect==4.9.0",
    "platformdirs==4.3.8",
    "plotly>=6.3.0",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "proto-plus==1.26.1",
    "protobuf==6.31.1",
    "psutil==7.0.0",
    "psycopg2-binary>=2.9.10",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings>=2.10.1",
    "pygments==2.19.2",
    "pymupdf>=1.26.3",
    "pytesseract>=0.3.13",
    "python-dateutil==2.9.0.post0",
    "python-json-logger==3.3.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "pyzmq==27.0.0",
    "referencing==0.36.2",
    "requests==2.32.4",
    "rfc3339-validator==0.1.4",
    "rfc3986-validator==0.1.1",
    "rpds-py==0.26.0",
    "rsa==4.9.1",
    "send2trash==1.8.3",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy>=2.0.43",
    "stack-data==0.6.3",
    "tenacity==8.5.0",
    "terminado==0.18.1",
    "tinycss2==1.4.0",
    "tornado==6.5.1",
    "traitlets==5.14.3",
    "types-python-dateutil==2.9.0.20250708",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uri-template==1.3.0",
    "urllib3==2.5.0",
    "wcwidth==0.2.13",
    "webcolors==24.11.1",
    "webencodings==0.5.1",
    "websocket-client==1.8.0",
    "websockets<14.0.0",
]
